requirements.txt
fastapi
uvicorn
python-docx
docxtpl
asyncpg
pydantic
docx2pdf